Changes to fix run issues
The first issue that was adding in a function to the CustomAdapter class for removed items.  The mutable list was then adjusted to notify of changes with the removed item. Then the delete button clicker was adjusted with an if statement for when the position changes and the count of the customAdapter changes too, then the new count changes the selection of the spinner